TRI-76 사용자 예약정보 검증 로직 구현 및 경로 수정
